+++ src/main/java/com/pyr/permission/domain/role/service/SysRoleUserService.java
@@ -21,37 +21,37 @@ public class SysRoleUserService {
     @Resource
     private SysUserMapper sysUserMapper;
 
-    private static boolean userIdsNotChange(List<Integer> userIds, List<Integer> originUserIdList) {
+    private static boolean userIdsNotChange(List<Long> userIds, List<Long> originUserIdList) {
         if (originUserIdList.size() == userIds.size()) {
-            Set<Integer> originUserIdSet = Sets.newHashSet(originUserIdList);
-            Set<Integer> userIdSet = Sets.newHashSet(userIds);
+            Set<Long> originUserIdSet = Sets.newHashSet(originUserIdList);
+            Set<Long> userIdSet = Sets.newHashSet(userIds);
             originUserIdSet.removeAll(userIdSet);
             return CollectionUtils.isEmpty(originUserIdSet);
         }
         return false;
     }
 
-    public List<SysUser> getListByRoleId(int roleId) {
-        List<Integer> userIdList = sysRoleUserMapper.getUserIdListByRoleId(roleId);
+    public List<SysUser> getListByRoleId(Long roleId) {
+        List<Long> userIdList = sysRoleUserMapper.getUserIdListByRoleId(roleId);
         if (CollectionUtils.isEmpty(userIdList)) {
             return Lists.newArrayList();
         }
         return sysUserMapper.getByIdList(userIdList);
     }
 
-    public void changeRoleUsers(int roleId, List<Integer> userIds) {
-        List<Integer> originUserIdList = sysRoleUserMapper.getUserIdListByRoleId(roleId);
+    public void changeRoleUsers(long roleId, List<Long> userIds) {
+        List<Long> originUserIdList = sysRoleUserMapper.getUserIdListByRoleId(roleId);
         if (userIdsNotChange(userIds, originUserIdList)) return;
         updateRoleUsers(roleId, userIds);
     }
 
-    private void updateRoleUsers(int roleId, List<Integer> userIdList) {
+    private void updateRoleUsers(long roleId, List<Long> userIdList) {
         sysRoleUserMapper.deleteByRoleId(roleId);
         if (CollectionUtils.isEmpty(userIdList)) {
             return;
         }
         List<SysRoleUser> roleUserList = Lists.newArrayList();
-        for (Integer userId : userIdList) {
+        for (Long userId : userIdList) {
             SysRoleUser roleUser = SysRoleUser.builder().roleId(roleId).userId(userId).build();
             roleUserList.add(roleUser);
         }
